.ipynb_checkpoints/.gitignore/Image-enhancement/requirements.txt
asgiref==3.6.0
contourpy==1.0.7
cycler==0.11.0
Django==4.1.5
fonttools==4.38.0
image==1.5.33
imageio==2.25.0
imutils==0.5.4
kiwisolver==1.4.4
matplotlib==3.6.3
networkx==3.0
numpy==1.24.1
opencv-python==4.7.0.68
packaging==23.0
Pillow==9.4.0
pyparsing==3.0.9
pytesseract==0.3.10
python-dateutil==2.8.2
PyWavelets==1.4.1
scikit-image==0.19.3
scipy==1.10.0
six==1.16.0
sqlparse==0.4.3
tifffile==2023.1.23.1
